feat: only change wallpaper iff the wallpaper actually changed

+++ .config/hypr/wallpaper.sh
@@ -5,20 +5,21 @@ menu() {
     find "${WALLPAPER_DIR}" -type f \( -iname "*.jpg" -o -iname "*.jpeg" -o -iname "*.png" -o -iname "*.gif" \) | awk '{print "img:"$0}'
 }
 main() {
+    cat $HOME/.cache/wal/colors-kitty.conf > /tmp/prev_wallpaper
     choice=$(menu | wofi -c ~/.config/wofi/wallpaper -s ~/.config/wofi/style-wallpaper.css --show dmenu --prompt "Select Wallpaper:" -n)
     selected_wallpaper=$(echo "$choice" | sed 's/^img://')
-    swww img "$selected_wallpaper" --transition-type any --transition-fps 60 --transition-duration .5
     wal -i "$selected_wallpaper" -n
-    swaync-client --reload-css
-    cat ~/.cache/wal/colors-kitty.conf > ~/.config/kitty/current-theme.conf
-    pywalfox update
-    color1=$(awk 'match($0, /color2=\47(.*)\47/,a) { print a[1] }' ~/.cache/wal/colors.sh)
-    color2=$(awk 'match($0, /color3=\47(.*)\47/,a) { print a[1] }' ~/.cache/wal/colors.sh)
-    cava_config="$HOME/.config/cava/config"
-    sed -i "s/^gradient_color_1 = .*/gradient_color_1 = '$color1'/" $cava_config
-    sed -i "s/^gradient_color_2 = .*/gradient_color_2 = '$color2'/" $cava_config
-    pkill -USR2 cava 2>/dev/null
-    source ~/.cache/wal/colors.sh && cp -r $wallpaper ~/wallpapers/pywallpaper.jpg 
+    cat $HOME/.cache/wal/colors-kitty.conf > /tmp/current_wallpaper
+    DIFF=$(diff /tmp/current_wallpaper /tmp/prev_wallpaper)
+    if [[ "$DIFF" != "" ]]; then
+      swww img "$selected_wallpaper" --transition-type any --transition-fps 60 --transition-duration .5
+      swaync-client --reload-css
+      cat ~/.cache/wal/colors-kitty.conf > ~/.config/kitty/current-theme.conf
+      pkill -USR2 cava 2>/dev/null
+      killall waybar
+      waybar &
+      source ~/.cache/wal/colors.sh && cp -r $wallpaper ~/wallpapers/pywallpaper.jpg 
+    fi
 }
 main
